Migrate from private GitLab to public GitHub
That way, we can use this package from Packagist in all our projects
without configuring GitLab. Engineers that got used to this coding
standard could also use it in their personal projects if they want.

That also implies that we would change the license. MIT is the most used
license on open-source projects and should work well for this library.

Signed-off-by: Henrique Moody <henrique.moody@instapro.com>